requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.130
botocore==1.29.130
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
discord.py==2.2.3
exceptiongroup==1.1.3
exchange-calendars==4.5
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lxml==4.9.3
matplotlib==3.8.0
mplfinance==0.12.10b0
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.0
outcome==1.2.0
packaging==23.2
pandas==2.1.1
pandas-market-calendars==4.3.1
peewee==3.16.3
Pillow==10.0.1
pycparser==2.21
pyee==8.2.2
pyluach==2.2.0
pyparsing==3.1.1
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
toolz==0.12.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.15
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.31
zipp==3.17.0
